feat(endpoint): address PR comments in new endpoint v2 pkgs
