+++ packages/middleware-endpoint/src/endpointMiddleware.ts
@@ -1,9 +1,7 @@
 import {
   MetadataBearer,
-  Pluggable,
   SerializeHandler,
   SerializeHandlerArguments,
-  SerializeHandlerOptions,
   SerializeHandlerOutput,
   SerializeMiddleware,
 } from "@aws-sdk/types";
@@ -24,24 +22,3 @@ export const endpointMiddleware = (options: {
     };
 };
 
-export const endpointMiddlewareOptions: SerializeHandlerOptions = {
-  step: "serialize",
-  tags: ["ENDPOINT_PARAMETERS", "ENDPOINT_V2", "ENDPOINT"],
-  name: "endpointMiddleware",
-  override: true,
-};
-
-export const getEndpointPlugin = (
-  config: any, //TODO(endpointsV2): should be ResolvedEndpointConfig interface
-  instruction: EndpointParameterInstruction
-): Pluggable<any, any> => ({
-  applyToStack: (clientStack) => {
-    clientStack.add(
-      endpointMiddleware({
-        config,
-        instruction,
-      }),
-      endpointMiddlewareOptions
-    );
-  },
-});
