+++ packages/middleware-endpoint/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./endpointMiddleware";
+export * from "./getEndpointPlugin";
 export * from "./resolveEndpointConfig";
 export * from "./types";
