+++ packages/types/src/index.ts
@@ -2,6 +2,7 @@ export * from "./abort";
 export * from "./auth";
 export * from "./client";
 export * from "./command";
+export * from "./credentials";
 export * from "./crypto";
 export * from "./endpoint";
 export * from "./eventStream";
